Pulls the template out to a resource package
