+++ packages/tools/adr_agent_tools/mcp_server/server.py
@@ -1,8 +1,11 @@
-import re
-from pathlib import Path
+from logging import Logger
 
 from mcp.server.fastmcp import Context, FastMCP
-from pydantic import BaseModel, Field
+from pydantic import AnyUrl
+
+from ..adr import AdrData, calculate_next_adr_number, read_adr_template, setup_adr_directory
+
+log = Logger(__name__)
 
 mcp = FastMCP(
     name="Adr",
@@ -11,71 +14,11 @@ mcp = FastMCP(
 )
 
 
-TEMPLATE = """
-# {short title, representative of solved problem and found solution}
-
-## Context and Problem Statement
-
-{Describe the context and problem statement, e.g., in free form using two to three sentences or in the form of an illustrative story. You may want to articulate the problem in form of a question and add links to collaboration boards or issue management systems.}
-
-## Considered Options
-
-* {title of option 1}
-* {title of option 2}
-* {title of option 3}
-* … <!-- numbers of options can vary -->
-
-## Decision Outcome
-
-Chosen option: "{title of option 1}", because {justification. e.g., only option, which meets k.o. criterion decision driver | which resolves force {force} | … | comes out best (see below)}.
-
-<!-- This is an optional element. Feel free to remove. -->
-### Consequences
-
-* Good, because {positive consequence, e.g., improvement of one or more desired qualities, …}
-* Bad, because {negative consequence, e.g., compromising one or more desired qualities, …}
-* … <!-- numbers of consequences can vary -->
-"""  # noqa: E501
-
-
-def setup_adr_directory(start_path: Path | str = "."):
-    """Find workspace root and ensure docs/adr directory exists."""
-    current = Path(start_path).resolve()
-    while current != current.parent:
-        if (current / ".git").exists():
-            adr_path = current / "docs" / "adr"
-            adr_path.mkdir(parents=True, exist_ok=True)
-            return adr_path
-        current = current.parent
-    raise FileNotFoundError("Git repository root not found")
-
-
-def calculate_next_adr_number(adr_directory: Path) -> str:
-    """Calculate the next ADR number based on existing files."""
-    pattern = re.compile(r"^(\d{4})-.*\.md$")
-    numbers = []
-
-    for file_path in adr_directory.glob("*.md"):
-        match = pattern.match(file_path.name)
-        if match:
-            numbers.append(int(match.group(1)))
-
-    next_number = max(numbers, default=0) + 1
-    return f"{next_number:04d}"
-
-
-class AdrData(BaseModel):
-    number_prefix: str = Field(
-        description="The number prefix for the ADR document (e.g., '0001')"
-    )
-    template: str = Field(
-        description="The ADR template content to be used for generating the document"
-    )
-
-
-@mcp.tool(description="Generates an ADR based on the context of the conversation",)
+@mcp.tool(
+    description="Generates an ADR based on the context of the conversation",
+)
 async def adr(cwd: str, ctx: Context) -> AdrData:
-    """Generate an ADR
+    """Generate metadata for creating an ADR(Architectural Decision Record)
 
     Args:
         cwd: Current working directory to start searching for git repository root
@@ -84,13 +27,20 @@ async def adr(cwd: str, ctx: Context) -> AdrData:
         AdrData: A package with instructions for the ADR that should be generated
     """
 
-    await ctx.info('Oh snap my neezy, we really in here')
-    print ('HEEEEEEEEEEEEEERRRRRRRRRRRRRE')
-    ctx.s
-
-
+    template_content = await ctx.read_resource(AnyUrl("template://MADR_MINIMAL_TEMPLATE"))
+    content = template_content[0]  # noqa: E501
+    template = content.content if hasattr(content, "content") else str(content)
 
     adr_directory = setup_adr_directory(cwd)
     number_prefix = calculate_next_adr_number(adr_directory)
-    res = AdrData(number_prefix=number_prefix, template=TEMPLATE)
+    res = AdrData(number_prefix=number_prefix, adr_directory=str(adr_directory), template=template)
     return res
+
+
+@mcp.resource("template://MADR_MINIMAL_TEMPLATE")
+def read_template() -> str:
+    return read_adr_template()
+
+
+if __name__ == "__main__":
+    mcp.run()
